passthrough response type
If the response type is PassThrough, the test server acts as a proxy.
This is not bullet-proof, with use of e.g. asyncio.gather, things can
mess up (but it was a problem even with the code before this commit).

From the remote server's headers, only the "Content-Type" is preserved
(which is enough in many cases). Other headers are filtered out because
may cause problems (for example content-encoding).

The PassThrough supports HTTPS. The support is maybe not very clean but it
works. The information about HTTPS usage in the client's request is
added to the request as a header, then the HTTPS is suppressed
(the request is converted to HTTP, as it was before this commit). The
proxy server reads the HTTPS information from the headers and modifies
the request's scheme according to it.