fix: allow mock https subrequest mocks. aiohttp_client compatibility
The `is_ssl` method was monkey-patched in a synchronous way that couldn't work when simultaneous requests needed it both patched and not-patched.  Instead of undoing the ClientRequest monkey-patching we create DirectClientRequest and have our passthrough session use it instead.

For compatibility between aresponses and aiohttp_client fixtures we use a work-around recommended by @serjshevchenko that replaces the aiohttp_client loop fixture with the event loop provided by pytest-asycnio.  The hanging was caused by different fixtures running on different event loops.