fix: restore the constructor to prevent RecursionError at 1000+ tests (#63)
When the fixture aresponses is used in more than 1000 consecutive tests (empirically, 940+), it hits the RecursionError when calling its ClientRequest's constructor. All tests that follow also fail regardless of what and how they test.

Caused by the constructor function being replaced, calling the original pre-replacement function, and never restoring the original function. As a result, every test adds +1 to the recursion, until it reaches some limits where Python raises a RecursionError for all frames (~995 frames from the entry point for a pure-Python script, and roughly ~940 level within pytest's context).

For the testing purposes, a limit is artificially lowered to 100 (of which 40-50 levels are taken by pytest & co, 10 for aresponses, and nearly 40 are consumed by recursion). The default 1000 is too slow (several seconds, not instant enough). Besides, it can vary in the future.